Improvements in components (simplifying)
